# docs/swagger.yaml
definitions:
  incoming.CreateUserParam:
    properties:
      password:
        type: string
      userName:
        type: string
    type: object
  model.Task:
    properties:
      assignName:
        type: string
      assignTo:
        type: string
      comment:
        type: string
      createTime:
        type: string
      desciption:
        type: string
      endTime:
        type: string
      id:
        type: string
      name:
        type: string
      priority:
        type: string
      sprint:
        type: string
      startTime:
        type: string
      status:
        type: string
      type:
        type: string
      userId:
        type: string
    type: object
  outgoing.ModelReturn:
    properties:
      userId:
        example: ""
        type: string
    type: object
  outgoing.UserReturn:
    properties:
      userId:
        example: "1234"
        type: string
    type: object
host: https://android-api
info:
  contact: {}
  description: This is a sample server for Android API.
  title: Android API
  version: "2.0"
paths:
  /task:
    post:
      parameters:
      - description: json of a task
        in: body
        name: data
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/outgoing.ModelReturn'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/outgoing.ModelReturn'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/outgoing.ModelReturn'
      summary: Create a task
      tags:
      - Task
    put:
      parameters:
      - description: json of a task
        in: body
        name: data
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/outgoing.ModelReturn'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/outgoing.ModelReturn'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/outgoing.ModelReturn'
      summary: Update a task
      tags:
      - Task
  /task/{id}:
    delete:
      parameters:
      - description: task id
        example: "1101"
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/outgoing.ModelReturn'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/outgoing.ModelReturn'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/outgoing.ModelReturn'
      summary: Delete a task
      tags:
      - Task
    get:
      parameters:
      - description: task id
        example: "1101"
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/outgoing.ModelReturn'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/outgoing.ModelReturn'
      summary: Detail a task
      tags:
      - Task
  /task/list-task/{userId}:
    get:
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/outgoing.ModelReturn'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/outgoing.ModelReturn'
      summary: Get list task
      tags:
      - Task
  /user:
    post:
      parameters:
      - description: userName and password
        in: body
        name: userInfo
        required: true
        schema:
          $ref: '#/definitions/incoming.CreateUserParam'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/outgoing.UserReturn'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/outgoing.ModelReturn'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/outgoing.ModelReturn'
      summary: Create account
      tags:
      - User
  /user/login:
    post:
      parameters:
      - description: userName and password
        in: body
        name: userInfo
        required: true
        schema:
          $ref: '#/definitions/incoming.CreateUserParam'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/outgoing.UserReturn'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/outgoing.ModelReturn'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/outgoing.ModelReturn'
      summary: Login
      tags:
      - User
schemes:
- http
swagger: "2.0"